pyproject.toml
[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
)
'''

[tool.poetry]
name = "proxy-service"
version = "0.0.1"
description = ""
authors = [

]

[tool.poetry.dependencies]
python = "^3.11"
gunicorn = "^20.1.0"
pip = "^23.1.2"
pytz = "^2023.3"

# Useful libs
attrs = "^23.1.0"
requests = "^2.30.0"
beautifulsoup4 = "^4.12.2"
python-decouple = "^3.8"

# Django
Django = "^4.2.1"
flaky = "^3.7.0"
faker = "^19.11.0"
factory-boy = "^3.3.0"
pytest = "^7.4.2"
pytest-responses = "^0.5.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.6.0"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
pytest-deadfixtures = "^2.2.1"
